+++ src/ObservableEither.ts
@@ -1,23 +1,27 @@
 /**
  * @since 0.6.8
  */
-import * as E from 'fp-ts/lib/Either'
-import * as R from './Observable'
-import * as TE from 'fp-ts/lib/TaskEither'
 import { Alt2 } from 'fp-ts/lib/Alt'
+import { Applicative2 } from 'fp-ts/lib/Applicative'
+import { Apply2 } from 'fp-ts/lib/Apply'
 import { Bifunctor2 } from 'fp-ts/lib/Bifunctor'
-import { MonadThrow2 } from 'fp-ts/lib/MonadThrow'
+import * as E from 'fp-ts/lib/Either'
+import { getEitherM } from 'fp-ts/lib/EitherT'
+import { Functor2 } from 'fp-ts/lib/Functor'
 import { IO } from 'fp-ts/lib/IO'
 import { IOEither } from 'fp-ts/lib/IOEither'
 import { Monad2 } from 'fp-ts/lib/Monad'
-import { MonadObservable2 } from './MonadObservable'
-import { Observable } from 'rxjs'
-import { Task } from 'fp-ts/lib/Task'
-import { getEitherM } from 'fp-ts/lib/EitherT'
+import { MonadIO2 } from 'fp-ts/lib/MonadIO'
+import { MonadTask2 } from 'fp-ts/lib/MonadTask'
+import { MonadThrow2 } from 'fp-ts/lib/MonadThrow'
 import { pipe, pipeable } from 'fp-ts/lib/pipeable'
-import { Applicative2 } from 'fp-ts/lib/Applicative'
+import { Task } from 'fp-ts/lib/Task'
+import * as TE from 'fp-ts/lib/TaskEither'
+import { Observable } from 'rxjs'
+import { MonadObservable2 } from './MonadObservable'
+import * as R from './Observable'
 
-const T = getEitherM(R.observable)
+const T = getEitherM(R.Monad)
 
 declare module 'fp-ts/lib/HKT' {
   interface URItoKind2<E, A> {
@@ -131,23 +135,156 @@ export function orElse<E, A, M>(
 export const swap: <E, A>(ma: ObservableEither<E, A>) => ObservableEither<A, E> = T.swap
 
 /**
+ * @since 0.6.12
+ */
+export const of: Applicative2<URI>['of'] = right
+
+// -------------------------------------------------------------------------------------
+// instances
+// -------------------------------------------------------------------------------------
+
+/**
+ * @since 0.6.12
+ */
+export const Functor: Functor2<URI> = {
+  URI,
+  map: T.map
+}
+
+/**
+ * @since 0.6.12
+ */
+export const Apply: Apply2<URI> = {
+  URI,
+  map: T.map,
+  ap: T.ap
+}
+
+/**
+ * @since 0.6.12
+ */
+export const Applicative: Applicative2<URI> = {
+  URI,
+  map: T.map,
+  ap: T.ap,
+  of
+}
+
+/**
+ * @since 0.6.12
+ */
+export const Monad: Monad2<URI> = {
+  URI,
+  map: T.map,
+  ap: T.ap,
+  of,
+  chain: T.chain
+}
+
+/**
+ * @since 0.6.12
+ */
+export const Bifunctor: Bifunctor2<URI> = {
+  URI,
+  bimap: T.bimap,
+  mapLeft: T.mapLeft
+}
+
+/**
+ * @since 0.6.12
+ */
+export const Alt: Alt2<URI> = {
+  URI,
+  map: T.map,
+  alt: T.alt
+}
+
+/**
+ * @since 0.6.12
+ */
+export const fromIO: MonadTask2<URI>['fromIO'] = rightIO
+
+/**
+ * @since 0.6.12
+ */
+export const MonadIO: MonadIO2<URI> = {
+  URI,
+  map: T.map,
+  ap: T.ap,
+  of,
+  chain: T.chain,
+  fromIO
+}
+
+/**
+ * @since 0.6.12
+ */
+export const MonadTask: MonadTask2<URI> = {
+  URI,
+  map: T.map,
+  ap: T.ap,
+  of,
+  chain: T.chain,
+  fromIO,
+  fromTask
+}
+
+/**
+ * @since 0.6.12
+ */
+export const fromObservable: MonadObservable2<URI>['fromObservable'] = rightObservable
+
+/**
+ * @since 0.6.12
+ */
+export const MonadObservable: MonadObservable2<URI> = {
+  URI,
+  map: T.map,
+  ap: T.ap,
+  of,
+  chain: T.chain,
+  fromIO,
+  fromTask,
+  fromObservable
+}
+
+/**
+ * @since 0.6.12
+ */
+export const throwError: MonadThrow2<URI>['throwError'] = left
+
+/**
+ * @since 0.6.12
+ */
+export const MonadThrow: MonadThrow2<URI> = {
+  URI,
+  map: T.map,
+  ap: T.ap,
+  of,
+  chain: T.chain,
+  throwError
+}
+
+/**
+ * @deprecated
  * @since 0.6.8
  */
 export const observableEither: Monad2<URI> & Bifunctor2<URI> & Alt2<URI> & MonadObservable2<URI> & MonadThrow2<URI> = {
   URI,
   map: T.map,
-  of: T.of,
+  of,
   ap: T.ap,
   chain: T.chain,
   bimap: T.bimap,
   mapLeft: T.mapLeft,
   alt: T.alt,
-  fromObservable: rightObservable,
-  fromTask,
   fromIO: rightIO,
+  fromTask,
+  fromObservable: rightObservable,
   throwError: left
 }
 
+// tslint:disable-next-line: deprecation
 const { alt, ap, apFirst, apSecond, bimap, chain, chainFirst, flatten, map, mapLeft } = pipeable(observableEither)
 
 export {
@@ -192,11 +329,6 @@ export {
    */
   mapLeft
 }
-
-/**
- * @since 0.6.12
- */
-export const of: Applicative2<URI>['of'] = right
 
 // -------------------------------------------------------------------------------------
 // do notation